pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "telegram-context-search-bot"
version = "0.2.0"
description = "Lightweight Telegram bot that answers group questions via NotebookLM, deployed on a VPS"
requires-python = ">=3.11"
dependencies = [
  "aiogram>=3.13.1",
  "fastapi>=0.115.0",
  "uvicorn>=0.30.6",
  "httpx>=0.27.2",
  "socksio>=1.0.0",
  "aiohttp-socks>=0.10.1",
  "pydantic-settings>=2.5.2",
  "python-dotenv>=1.0.1",
  "orjson>=3.10.7",
  "notebooklm-py==0.3.4",
  "cryptography>=45.0.0",
  "Pillow>=10.4.0"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
